Update README.md
1-Project
1.1. Project Topic
Heart Disease Data Mining- Multi-Source Data Fusion and Analysis for Clinical. In this project, data mining methods related to heart disease were utilized to analyze data obtained from various clinical sources. Throughout this process, a dataset comprising the combination of five separate datasets containing different variables such as patients' age, gender, blood pressure tests, and more was employed. Subsequently, these data were scrutinized in terms of relationships and attributes that could be employed to identify and diagnose risk factors for heart disease.

1.2.Goal Of The Project
This project aims to promote the use of data mining and machine learning techniques for the early diagnosis and effective treatment of heart diseases. Heart disease is recognized as one of the leading causes of death worldwide, and early diagnosis and effective treatment can prevent disease progression or manage risk factors, thereby improving quality of life. Therefore, this project focuses on analyzing data obtained from various clinical sources to identify risk factors crucial for the determination, diagnosis, and treatment of heart disease.

1.3.The Success Criteria Of The Project
The extent to which the developed models can be integrated into clinical practice and the methods for achieving this integration, as well as the potential contribution of these models to diagnostic and treatment outcomes, may be regarded as definitive indicators of success. The success of this integration would reflect the potential of advanced technologies to be effectively utilized in enhancing the quality of healthcare services and improving patients' health outcomes.